move nested beans from ognl to a not excluded package
IntelliJ powerful refactoring has done it.
