+++ core/src/test/java/com/opensymphony/xwork2/beans/ognl/BadJavaBean.java
@@ -1,0 +1,40 @@
+/*
+ * Licensed to the Apache Software Foundation (ASF) under one
+ * or more contributor license agreements.  See the NOTICE file
+ * distributed with this work for additional information
+ * regarding copyright ownership.  The ASF licenses this file
+ * to you under the Apache License, Version 2.0 (the
+ * "License"); you may not use this file except in compliance
+ * with the License.  You may obtain a copy of the License at
+ *
+ *  http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing,
+ * software distributed under the License is distributed on an
+ * "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
+ * KIND, either express or implied.  See the License for the
+ * specific language governing permissions and limitations
+ * under the License.
+ */
+package com.opensymphony.xwork2.beans.ognl;
+
+public class BadJavaBean {
+    private int count;
+    private int count2;
+
+    public void setCount(int count) {
+        this.count = count;
+    }
+
+    public String getCount() {
+        return "" + count;
+    }
+
+    public void setCount2(String count2) {
+        this.count2 = Integer.parseInt(count2);
+    }
+
+    public int getCount2() {
+        return count2;
+    }
+}
